src/requirements.txt
# Django requirements
django==1.10.5 # Django
pyyaml==3.12 # Used for fixtures
pillow==4.0.0 # For image uploading
django-gravatar2==1.4.0 # For gravatar
django-cleanup==0.4.2 # Remove images when replacing them
django-ganalytics==0.4 # Include analytics from settings
django-widget-tweaks==1.4.1 # For editing form attributes in templates
django-markdown-deux==1.0.5 # For rendering markdown
jsonfield==1.0.3 # For JSON serialization
raven==5.32.0 # for Sentry error reporting
git+https://github.com/djangonauts/django-pgjson.git # Using Postgres JSON field (can't use PIP as it's out of date)
python-memcached==1.58 # Memcached bindings
git+https://github.com/rhec/django-html5.git
django-ical==1.4 # For ical feed generation
django-queryset-csv==1.0.0
djangorestframework==3.5.3
markdown==2.6.7
django-filter==1.0.1
django-wkhtmltopdf==3.1.0
django-ordered-model==1.3.0
django-guardian==1.4.6


# Heroku
gunicorn==19.6.0 # For hosting on heroku
dj-database-url==0.4.2 # Heroku databases
dj-static==0.0.6 # Static files
psycopg2==2.6.2 # Postgres driver
django-storages==1.5.1 # Media storage on AWS
boto==2.45.0 # Media storage on AWS

# Authentication

django-allauth==0.30.0 # Authentication and social auth

# General

pytz==2016.10 # To stop python moaning about naive timezones
cached-property==1.3.0 # Add cachable properties
vcrpy==1.10.5 # For mocking HTTP requests
python-dateutil==2.6.0 # For date manipulations
stripe==1.46.0 # For accepting payment
beautifulsoup4==4.5.3 # For parsing HTML
python-dotenv==0.6.2 # To use .env files
gitpython==2.1.1 # For downloading plugins
celery==4.0.2 # For periodic and background tasks
redis==2.10.5 # For periodic and background tasks
